+++ lib/metaApi/metaApiConnection.py
@@ -10,16 +10,19 @@ import asyncio
 class MetaApiConnection(SynchronizationListener, ReconnectListener):
     """Exposes MetaApi MetaTrader API connection to consumers."""
 
-    def __init__(self, websocket_client: MetaApiWebsocketClient, account: MetatraderAccountModel):
+    def __init__(self, websocket_client: MetaApiWebsocketClient, account: MetatraderAccountModel,
+                 application: str = None):
         """Inits MetaApi MetaTrader Api connection.
 
         Args:
             websocket_client: MetaApi websocket client.
             account: MetaTrader account id to connect to.
+            application: Application to use.
         """
         super().__init__()
         self._websocketClient = websocket_client
         self._account = account
+        self._application = application
 
     def create_market_buy_order(self, symbol: str, volume: float, stop_loss: float or StopOptions = None,
                                 take_profit: float or StopOptions = None,
@@ -43,7 +46,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         trade_params = {'actionType': 'ORDER_TYPE_BUY', 'symbol': symbol, 'volume': volume,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_market_sell_order(self, symbol: str, volume: float, stop_loss: float or StopOptions = None,
                                  take_profit: float or StopOptions = None,
@@ -67,7 +70,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         trade_params = {'actionType': 'ORDER_TYPE_SELL', 'symbol': symbol, 'volume': volume,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_limit_buy_order(self, symbol: str, volume: float, open_price: float,
                                stop_loss: float or StopOptions = None,
@@ -93,7 +96,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_limit_sell_order(self, symbol: str, volume: float, open_price: float,
                                 stop_loss: float or StopOptions = None,
@@ -119,7 +122,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_stop_buy_order(self, symbol: str, volume: float, open_price: float,
                               stop_loss: float or StopOptions = None,
@@ -145,7 +148,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_stop_sell_order(self, symbol: str, volume: float, open_price: float,
                                stop_loss: float or StopOptions = None,
@@ -171,7 +174,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_stop_limit_buy_order(self, symbol: str, volume: float, open_price: float, stop_limit_price: float,
                                     stop_loss: float or StopOptions = None, take_profit: float or StopOptions = None,
@@ -197,7 +200,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price, 'stopLimitPrice': stop_limit_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def create_stop_limit_sell_order(self, symbol: str, volume: float, open_price: float, stop_limit_price: float,
                                      stop_loss: float or StopOptions = None, take_profit: float or StopOptions = None,
@@ -223,7 +226,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
                         'openPrice': open_price, 'stopLimitPrice': stop_limit_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def modify_position(self, position_id: str, stop_loss: float or StopOptions = None,
                         take_profit: float or StopOptions = None) -> \
@@ -243,7 +246,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         """
         trade_params = {'actionType': 'POSITION_MODIFY', 'positionId': position_id,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def close_position_partially(self, position_id: str, volume: float, options: MarketTradeOptions = None) -> \
             'Coroutine[asyncio.Future[MetatraderTradeResponse]]':
@@ -262,7 +265,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         """
         trade_params = {'actionType': 'POSITION_PARTIAL', 'positionId': position_id, 'volume': volume}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def close_position(self, position_id: str, options: MarketTradeOptions = None) -> \
             'Coroutine[asyncio.Future[MetatraderTradeResponse]]':
@@ -280,7 +283,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         """
         trade_params = {'actionType': 'POSITION_CLOSE_ID', 'positionId': position_id}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def close_by(self, position_id: str, opposite_position_id: str, options: MarketTradeOptions = None) -> \
             'Coroutine[asyncio.Future[MetatraderTradeResponse]]':
@@ -300,7 +303,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         trade_params = {'actionType': 'POSITION_CLOSE_BY', 'positionId': position_id,
                         'closeByPositionId': opposite_position_id}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def close_positions_by_symbol(self, symbol: str, options: MarketTradeOptions = None) -> \
             'Coroutine[asyncio.Future[MetatraderTradeResponse]]':
@@ -318,7 +321,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         """
         trade_params = {'actionType': 'POSITIONS_CLOSE_SYMBOL', 'symbol': symbol}
         trade_params.update(options or {})
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def modify_order(self, order_id: str, open_price: float, stop_loss: float or StopOptions = None,
                      take_profit: float or StopOptions = None) -> \
@@ -339,7 +342,7 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         """
         trade_params = {'actionType': 'ORDER_MODIFY', 'orderId': order_id, 'openPrice': open_price,
                         **self._generate_stop_options(stop_loss=stop_loss, take_profit=take_profit)}
-        return self._websocketClient.trade(self._account.id, trade_params)
+        return self._websocketClient.trade(self._account.id, trade_params, self._application)
 
     def cancel_order(self, order_id: str) -> \
             'Coroutine[asyncio.Future[MetatraderTradeResponse]]':
@@ -354,7 +357,8 @@ class MetaApiConnection(SynchronizationListener, ReconnectListener):
         Raises:
             TradeException: On trade error, check error properties for error code details.
         """
-        return self._websocketClient.trade(self._account.id, {'actionType': 'ORDER_CANCEL', 'orderId': order_id})
+        return self._websocketClient.trade(self._account.id, {'actionType': 'ORDER_CANCEL', 'orderId': order_id},
+                                           self._application)
 
     def reconnect(self) -> Coroutine:
         """Reconnects to the Metatrader terminal (see https://metaapi.cloud/docs/client/websocket/api/reconnect/).