+++ changelog.md
@@ -1,3 +1,8 @@
+15.1.0
+  - added option to keep server-side subscription when retrieving latest market data via RPC API
+  - updated CopyFactory API to 3.1.0
+  - fixed rpc connection requests
+
 15.0.1
   - fix symbol price model
   - fix terminal state hashing